Moved misc_tools from testing to toxcore.
