+++ other/Makefile.inc
@@ -3,8 +3,7 @@ noinst_PROGRAMS += DHT_bootstrap
 DHT_bootstrap_SOURCES = $(top_srcdir)/other/DHT_bootstrap.c \
                         $(top_srcdir)/toxcore/DHT.h \
                         $(top_srcdir)/toxcore/friend_requests.h \
-                        $(top_srcdir)/testing/misc_tools.h \
-                        $(top_srcdir)/testing/misc_tools.c
+                        $(top_srcdir)/toxcore/misc_tools.h
 
 DHT_bootstrap_CFLAGS = -I$(top_srcdir)/other \
                         $(LIBSODIUM_CFLAGS)
