+++ testing/nTox.c
@@ -21,7 +21,7 @@
  *
  */
 #include "nTox.h"
-#include "misc_tools.h"
+#include "../toxcore/misc_tools.h"
 
 #include <stdio.h>
 #include <time.h>
