+++ toxcore/misc_tools.c
