+++ toxcore/misc_tools.h
